musicbot/requirements.txt
discord.py[voice]
wavelink
yt-dlp